feat: init layout with sidebar with page qualification action
