+++ app/(connected)/(with-sidebar)/(without-chat)/evenements/[id_evenement]/cloture/page.tsx
@@ -3,7 +3,7 @@ import { notFound, redirect } from 'next/navigation'
 
 import CloturePage from 'app/(connected)/(with-sidebar)/(without-chat)/evenements/[id_evenement]/cloture/CloturePage'
 import PageHeaderPortal from 'components/PageHeaderPortal'
-import PageRetourPortal from 'components/PageRetourPortal'
+import { PageRetourPortal } from 'components/PageNavigationPortals'
 import { estUserPoleEmploi } from 'interfaces/conseiller'
 import { StatutAnimationCollective } from 'interfaces/evenement'
 import { getDetailsEvenement } from 'services/evenements.service'
