+++ components/FilAriane.tsx
@@ -4,50 +4,48 @@ import IconComponent, { IconName } from 'components/ui/IconComponent'
 import mapRoutesToLabels from 'utils/route-labels'
 
 type FilArianeProps = {
-  currentPath: string
+  path: string
 }
 
-export default function FilAriane({ currentPath }: FilArianeProps) {
-  const ariane = creationFilAriane(currentPath)
+export default function FilAriane({ path }: FilArianeProps) {
+  const ariane = creationFilAriane(path)
 
   if (!ariane.length) {
     return null
   }
 
   return (
-    <nav aria-label="Fil d'ariane">
-      <ol className='mb-2 flex items-center'>
-        {ariane.map(({ href, label }, index) => (
-          <li key={label} className='flex items-center'>
-            {index < ariane.length - 1 && (
-              <>
-                <Link
-                  href={href}
-                  className='text-s-regular text-content_color underline hover:text-primary_darken'
-                >
-                  {label}
-                </Link>
-                <IconComponent
-                  name={IconName.ChevronRight}
-                  aria-hidden={true}
-                  focusable={false}
-                  className='mx-2 w-6 h-6 fill-content_color'
-                />
-              </>
-            )}
-            {index === ariane.length - 1 && (
+    <ol className='mb-2 flex items-center'>
+      {ariane.map(({ href, label }, index) => (
+        <li key={label} className='flex items-center'>
+          {index < ariane.length - 1 && (
+            <>
               <Link
                 href={href}
-                aria-current='page'
-                className='text-s-regular text-content_color'
+                className='text-s-regular text-content_color underline hover:text-primary_darken'
               >
                 {label}
               </Link>
-            )}
-          </li>
-        ))}
-      </ol>
-    </nav>
+              <IconComponent
+                name={IconName.ChevronRight}
+                aria-hidden={true}
+                focusable={false}
+                className='mx-2 w-6 h-6 fill-content_color'
+              />
+            </>
+          )}
+          {index === ariane.length - 1 && (
+            <Link
+              href={href}
+              aria-current='page'
+              className='text-s-regular text-content_color'
+            >
+              {label}
+            </Link>
+          )}
+        </li>
+      ))}
+    </ol>
   )
 }
