--- tests/pages/Qualification.page.test.tsx
@@ -1,373 +1,0 @@
-import { fireEvent, screen, within } from '@testing-library/react'
-import userEvent from '@testing-library/user-event'
-import { DateTime } from 'luxon'
-import { GetServerSidePropsResult } from 'next'
-import { useRouter } from 'next/router'
-import { GetServerSidePropsContext } from 'next/types'
-
-import { desSituationsNonProfessionnelles, uneAction } from 'fixtures/action'
-import {
-  Action,
-  SituationNonProfessionnelle,
-  StatutAction,
-} from 'interfaces/action'
-import PageQualification, {
-  getServerSideProps,
-} from 'pages/mes-jeunes/[jeune_id]/actions/[action_id]/qualification'
-import { AlerteParam } from 'referentiel/alerteParam'
-import {
-  getAction,
-  getSituationsNonProfessionnelles,
-  qualifier,
-} from 'services/actions.service'
-import renderWithContexts from 'tests/renderWithContexts'
-import withMandatorySessionOrRedirect from 'utils/auth/withMandatorySessionOrRedirect'
-
-jest.mock('utils/auth/withMandatorySessionOrRedirect')
-jest.mock('services/actions.service')
-jest.mock('next/router')
-
-describe("Page Qualification d'une action", () => {
-  describe('server side', () => {
-    it('requiert une session valide', async () => {
-      // Given
-      ;(withMandatorySessionOrRedirect as jest.Mock).mockResolvedValue({
-        validSession: false,
-        redirect: 'wherever',
-      })
-
-      // When
-      const actual = await getServerSideProps({} as GetServerSidePropsContext)
-
-      // Then
-      expect(withMandatorySessionOrRedirect).toHaveBeenCalled()
-      expect(actual).toEqual({ redirect: 'wherever' })
-    })
-
-    describe("quand le conseiller n'est pas MiLo", () => {
-      it('renvoie une 404', async () => {
-        // Given
-        ;(withMandatorySessionOrRedirect as jest.Mock).mockResolvedValue({
-          validSession: true,
-          session: {
-            user: { structure: 'POLE_EMPLOI' },
-          },
-        })
-
-        // When
-        const actual = await getServerSideProps({} as GetServerSidePropsContext)
-
-        // Then
-        expect(withMandatorySessionOrRedirect).toHaveBeenCalled()
-        expect(actual).toEqual({ notFound: true })
-      })
-    })
-
-    describe('quand le conseiller est MiLo', () => {
-      beforeEach(() => {
-        ;(withMandatorySessionOrRedirect as jest.Mock).mockResolvedValue({
-          validSession: true,
-          session: {
-            accessToken: 'accessToken',
-            user: { structure: 'MILO' },
-          },
-        })
-      })
-
-      describe("quand l'action n'existe pas", () => {
-        it('renvoie une 404', async () => {
-          ;(getAction as jest.Mock).mockResolvedValue(undefined)
-
-          // When
-          const actual: GetServerSidePropsResult<any> =
-            await getServerSideProps({
-              query: { action_id: 'id-action' },
-            } as unknown as GetServerSidePropsContext)
-
-          // Then
-          expect(actual).toEqual({ notFound: true })
-        })
-      })
-
-      describe("quand l'action n'est pas terminée", () => {
-        it('renvoie une 404', async () => {
-          ;(getAction as jest.Mock).mockResolvedValue({
-            action: uneAction(),
-            jeune: {
-              id: 'jeune-1',
-              prenom: 'Nadia',
-              nom: 'Sanfamiye',
-            },
-          })
-
-          // When
-          const actual: GetServerSidePropsResult<any> =
-            await getServerSideProps({
-              query: { action_id: 'id-action' },
-            } as unknown as GetServerSidePropsContext)
-
-          // Then
-          expect(actual).toEqual({ notFound: true })
-        })
-      })
-
-      describe("quand l'action est qualifiée", () => {
-        it('renvoie une 404', async () => {
-          ;(getAction as jest.Mock).mockResolvedValue({
-            action: uneAction({
-              status: StatutAction.Terminee,
-              qualification: {
-                libelle: 'Santé',
-                isSituationNonProfessionnelle: true,
-              },
-            }),
-            jeune: {
-              id: 'jeune-1',
-              prenom: 'Nadia',
-              nom: 'Sanfamiye',
-            },
-          })
-
-          // When
-          const actual: GetServerSidePropsResult<any> =
-            await getServerSideProps({
-              query: { action_id: 'id-action' },
-            } as unknown as GetServerSidePropsContext)
-
-          // Then
-          expect(actual).toEqual({ notFound: true })
-        })
-      })
-
-      describe("quand l'action est terminée", () => {
-        it('récupère la liste des situations non professionnelles', async () => {
-          // Given
-          const action = uneAction({ status: StatutAction.Terminee })
-          const situationsNonProfessionnelles =
-            desSituationsNonProfessionnelles()
-          ;(getAction as jest.Mock).mockResolvedValue({
-            action,
-            jeune: {
-              id: 'jeune-1',
-              prenom: 'Nadia',
-              nom: 'Sanfamiye',
-            },
-          })
-          ;(getSituationsNonProfessionnelles as jest.Mock).mockResolvedValue(
-            situationsNonProfessionnelles
-          )
-
-          // When
-          const actual = await getServerSideProps({
-            query: { action_id: action.id },
-          } as unknown as GetServerSidePropsContext)
-
-          // Then
-          expect(getAction).toHaveBeenCalledWith(action.id, 'accessToken')
-          expect(getSituationsNonProfessionnelles).toHaveBeenCalledWith(
-            'accessToken'
-          )
-          expect(actual).toEqual({
-            props: {
-              action,
-              situationsNonProfessionnelles,
-              pageTitle: 'Actions jeune - Qualifier action',
-              pageHeader: 'Créer une situation non professionnelle',
-              returnTo: '/mes-jeunes/jeune-1/actions/id-action-1',
-              withoutChat: true,
-            },
-          })
-        })
-      })
-    })
-  })
-
-  describe('client side', () => {
-    let action: Action
-    let situationsNonProfessionnelles: SituationNonProfessionnelle[]
-
-    let alerteSetter: (key: AlerteParam | undefined, target?: string) => void
-    let push: jest.Mock
-    beforeEach(() => {
-      // Given
-      action = uneAction({ dateFinReelle: '2022-09-02T11:00:00.000Z' })
-      situationsNonProfessionnelles = desSituationsNonProfessionnelles()
-
-      alerteSetter = jest.fn()
-      push = jest.fn()
-      ;(useRouter as jest.Mock).mockReturnValue({ push })
-
-      // When
-      renderWithContexts(
-        <PageQualification
-          action={action}
-          situationsNonProfessionnelles={situationsNonProfessionnelles}
-          pageTitle=''
-          returnTo='/mes-jeunes/jeune-1/actions/id-action-1'
-        />,
-        {
-          customAlerte: { alerteSetter },
-        }
-      )
-    })
-
-    it('affiche un message d’information', async () => {
-      // Then
-      expect(
-        screen.getByText(
-          'Ces informations seront intégrées sur le dossier i-milo du bénéficiaire'
-        )
-      ).toBeInTheDocument()
-      expect(
-        screen.getByText(/respecter les Conditions Générales d’utilisation/)
-      ).toBeInTheDocument()
-      expect(
-        screen.getByRole('link', {
-          name: 'Voir le détail des CGU (nouvelle fenêtre)',
-        })
-      ).toHaveAttribute(
-        'href',
-        'https://c-milo.i-milo.fr/jcms/t482_1002488/fr/mentions-legales'
-      )
-      expect(
-        screen.getByRole('link', {
-          name: 'Voir le détail de l’arrêté du 17 novembre 2021 (nouvelle fenêtre)',
-        })
-      ).toHaveAttribute(
-        'href',
-        'https://www.legifrance.gouv.fr/jorf/id/JORFTEXT000045084361'
-      )
-    })
-
-    it("affiche le résumé de l'action", () => {
-      // Then
-      const etape1 = screen.getByRole('group', {
-        name: "Étape 1 Résumé de l'action",
-      })
-      expect(
-        within(etape1).getByRole('textbox', {
-          name: /Intitulé et description de l'action/,
-        })
-      ).toHaveValue(action.content + ' - ' + action.comment)
-    })
-
-    it('demande un type de situation non professionnelle', () => {
-      // Then
-      const etape2 = screen.getByRole('group', {
-        name: 'Étape 2 Type',
-      })
-      const selectSNP = within(etape2).getByRole('combobox', { name: 'Type' })
-      situationsNonProfessionnelles.forEach(({ code, label }) => {
-        expect(
-          within(selectSNP).getByRole('option', { name: label })
-        ).toHaveValue(code)
-      })
-    })
-
-    it("permet de modifier la date de début de l'action", () => {
-      // Then
-      const etape3 = screen.getByRole('group', {
-        name: 'Étape 3 Date de début de l’action',
-      })
-      const inputDate = within(etape3).getByLabelText('* Date de début')
-      expect(inputDate).toHaveAttribute('type', 'date')
-      expect(inputDate).toHaveValue('2022-02-15')
-    })
-
-    it("permet de modifier la date de fin réelle de l'action", () => {
-      // Then
-      const etape4 = screen.getByRole('group', {
-        name: 'Étape 4 Date de fin de l’action',
-      })
-      const inputDate = within(etape4).getByLabelText('* Date de fin')
-      expect(inputDate).toHaveAttribute('type', 'date')
-      expect(inputDate).toHaveAttribute('min', '2022-02-15')
-      expect(inputDate).toHaveValue('2022-09-02')
-    })
-
-    describe('validation formulaire', () => {
-      let inputCommentaire: HTMLElement
-      beforeEach(async () => {
-        // Given
-        inputCommentaire = screen.getByRole('textbox', {
-          name: /Intitulé et description/,
-        })
-        const selectSNP = screen.getByRole('combobox', { name: 'Type' })
-        const inputDate = screen.getByLabelText('* Date de fin')
-
-        await userEvent.clear(inputCommentaire)
-        await userEvent.type(inputCommentaire, 'Nouveau commentaire modifié')
-        await userEvent.selectOptions(
-          selectSNP,
-          situationsNonProfessionnelles[1].code
-        )
-        // FIXME userEvent.type ne marche pas bien avec les input date/time
-        fireEvent.change(inputDate, { target: { value: '2022-09-05' } })
-      })
-
-      it('envoie la qualification au fuseau horaire du navigateur du client', async () => {
-        // When
-        await userEvent.click(
-          screen.getByRole('button', { name: 'Créer et envoyer à i-milo' })
-        )
-
-        // Then
-        expect(qualifier).toHaveBeenCalledWith(action.id, 'SNP_2', {
-          commentaire: 'Nouveau commentaire modifié',
-          dateDebutModifiee: DateTime.fromISO(
-            '2022-02-15T00:00:00.000+01:00' // en février, l'offset est +1 (DST)
-          ),
-          dateFinModifiee: DateTime.fromISO('2022-09-05T00:00:00.000+02:00'),
-        })
-      })
-
-      it("redirige vers le détail de l'action", async () => {
-        // When
-        await userEvent.click(
-          screen.getByRole('button', { name: 'Créer et envoyer à i-milo' })
-        )
-
-        // Then
-        expect(alerteSetter).toHaveBeenCalledWith('qualificationSNP')
-        expect(push).toHaveBeenCalledWith(
-          '/mes-jeunes/jeune-1/actions/id-action-1'
-        )
-      })
-
-      it('est désactivée sur le commentaire n’est pas renseigné', async () => {
-        // When
-        await userEvent.clear(inputCommentaire)
-        await userEvent.tab()
-
-        // Then
-        expect(
-          screen.getByText(
-            'Le champ Intitulé et description n’est pas renseigné. Veuillez renseigner une description.'
-          )
-        ).toBeInTheDocument()
-        expect(screen.getByRole('button', { name: /Créer/ })).toHaveAttribute(
-          'disabled',
-          ''
-        )
-      })
-
-      it('est désactivée sur le commentaire contient plus de 255 caractères', async () => {
-        // When
-        await userEvent.clear(inputCommentaire)
-        await userEvent.type(inputCommentaire, 'a'.repeat(256))
-        await userEvent.tab()
-
-        // Then
-        expect(
-          screen.getByText(
-            'Vous avez dépassé le nombre maximal de caractères. Veuillez retirer des caractères.'
-          )
-        ).toBeInTheDocument()
-        expect(screen.getByRole('button', { name: /Créer/ })).toHaveAttribute(
-          'disabled',
-          ''
-        )
-      })
-    })
-  })
-})
